More on internships
Internships are saved in the DB in the schedule table
and are correctly produced on the blade.php.
Internships cannot be added to semesters with courses

deleting internships works that are loaded onto the blade.php
NOTE: deleting dynamically added internships isn't working yet.
Internship cards are un sortable and the semesters which they contain
are cant house new courses